Link to clients as directory for the sake of CF+B2
My VPS service is going away, and so instead of selfhosting, I am moving
my static site to Backblaze B2 + Cloudflare.

B2 does not do automatic URL rewriting from a directory to that
directory's index.html, and instead responds with a 404 for such URls.
To work around that, I am rewriting paths in a Cloudflare worker, which
matches on paths ending in '/'.

Update the link to my clients page to include a trailing slash, so that
the Cloudflare worker will translate that to 'clients/index.html'.

Related resources:
- https://help.backblaze.com/hc/en-us/articles/217666928-Using-Backblaze-B2-with-the-Cloudflare-CDN
- https://www.grahn.io/posts/2020-02-08-s3-vs-b2-static-web-hosting/
- https://github.com/ChrisFigura/worker-backblaze